Fix the regex used to extract words
Remove the html decoding/encoding of the string as it strips
new line information. We'll assume that splitTextToArray_FirstPass
should receive a string of text without any html entities in it.

In splitTextToArray_FirstPass, add \r\n and \n to the selected
characters returned by the regex in order to generate pauses
for new lines and paragraphs later in getTextArrayBasic.